Move getCategoryArticles to api.controller.js and remove category.controller.js
